refactor: update import paths for FeatureFlagKey
Replaced references to '~/generated-metadata/graphql' with '~/generated/graphql' across multiple files. This change ensures uniformity in import paths and aligns with the updated directory structure.
