+++ packages/twenty-front/src/modules/activities/components/ActivityRichTextEditor.tsx
@@ -31,7 +31,7 @@ import { useIsFeatureEnabled } from '@/workspace/hooks/useIsFeatureEnabled';
 import '@blocknote/core/fonts/inter.css';
 import '@blocknote/mantine/style.css';
 import '@blocknote/react/style.css';
-import { FeatureFlagKey } from '~/generated-metadata/graphql';
+import { FeatureFlagKey } from '~/generated/graphql';
 
 type ActivityRichTextEditorProps = {
   activityId: string;
